+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-4.0.7 (unreleased)
+4.0.7 (2024-02-20)
 ------------------
 
 - @bandi-search-filters: if called on a specific folder returns only tipologies and subjects of bandi inside that folder.